readme: make the guide even more simple
The guide for creating a UDF was missing a few steps
that would be obvious for a Rust developer, but if
a Rust newcomer tries to follow the guide, he might
get an error. When following the guide now, all
commands should finish successfully, even if the
user is performing them blindly.

We also mention the `wasm-strip` tool for for more
optional optimizations