CI: commit cargo lockfile for MSRV
It's good to test the library against the newest version of the
dependencies, but this is not really viable for the minimum supported
rust version check - some dependencies reserve the right to bump their
MSRV in semver-compatible releases. In fact, the newest version of tokio
(1.14.0) no longer supports rust 1.66.1.

Commit a lockfile which specifies exact dependency versions which are
known to work with rust 1.66.1. The lockfile was created by letting
cargo check construct a new lockfile and then pinning the tokio version
to 1.38.0.

The lockfile is committed as Cargo.lock.msrv and renamed to Cargo.lock
only during the MSRV CI job, so this does not interfere with the regular
developer workflow and the non-MSRV build job.